Finish HTML shell & Add images for info cards
Finish basic shell of the HTML document for the skeleton. Almost all
elements should be in the doc at this point. Some basic flexbox has
been added to start help visualizing the project as it coalesces. I am
going to try and take it as 1-for-1 as possible.

Beyond that, I uploaded images from my personal phone. They are all
mine and therefore safe to use without infringing on copyright.